+++ programs/voter-stake-registry/src/governance.rs
@@ -61,8 +61,7 @@ macro_rules! vote_weight_record {
         impl anchor_lang::Discriminator for VoterWeightRecord {
             const DISCRIMINATOR: [u8; 8] = [0; 8];
             fn discriminator() -> [u8; 8] {
-                // legacy discriminator which is not compatible with Anchor but is used by older plugins
-                *b"2ef99b4b"
+                spl_governance_addin_api::voter_weight::VoterWeightRecord::ACCOUNT_DISCRIMINATOR
             }
         }
     };