+++ programs/voter-stake-registry/tests/test_voting.rs
@@ -6,7 +6,7 @@ use voter_stake_registry::state::LockupKind;
 
 mod program_test;
 
-#[allow(unaligned_references)]
+
 #[tokio::test]
 async fn test_voting() -> Result<(), TransportError> {
     let context = TestContext::new().await;
@@ -126,7 +126,7 @@ async fn test_voting() -> Result<(), TransportError> {
         )
         .await
         .expect_err("not enough tokens to create proposal");
-
+    
     addin
         .deposit(
             &registrar,