+++ run-release.sh
@@ -18,7 +18,7 @@ anchor build
 # update on chain program and IDL, atm used for testing/developing
 anchor deploy --provider.cluster devnet --provider.wallet ${PROVIDER_WALLET}
 anchor idl upgrade --provider.cluster devnet --provider.wallet ${PROVIDER_WALLET}\
- --filepath target/idl/voter_stake_registry.json 4Q6WW2ouZ6V3iaNm56MTd5n2tnTm4C5fiH8miFHnAFHo
+ --filepath target/idl/voter_stake_registry.json 9SJqwCQ5AJkFtC7zxfFsF6Y5dm22XzN3JEhn3N14v23t
 
 # update types in npm package and publish the npm package
 cp ./target/types/voter_stake_registry.ts src/voter_stake_registry.ts